lastest

+++ app/assets/javascripts/managements/course_manage.js
@@ -32,8 +32,13 @@ courseManage = {
             items: [{
                 text: '删除',
                 scope: this,
+                iconCls: Wando.icons.deletes,
                 handler: function(){
-                    this.deleteCourse();
+                    Wando.msg.confirm('确定', "确定要删除？", function(btn){ 
+                        if(btn=='ok'){ 
+                            this.deleteCourse();
+                        }   
+                    }, this)
                 }
             }]
         });
@@ -67,12 +72,13 @@ courseManage = {
             items: [{
                 text: '添加',
                 scope: this,
+                iconCls: Wando.icons.add,
                 handler: function(){
                     var form = Ext.getCmp('courseForm').getForm();
                     if(form.isValid()){
                         this.addCourse();
                     }else{
-                        Ext.Msg.alert('警告', '请把数据填写完整');
+                        Wando.msg.warning('警告', '请输入完整、正确的数据类型的数据');
                     }
 
                 }
@@ -138,6 +144,8 @@ courseManage = {
                 store: this.createTStore()
             }, {
                 fieldLabel: '选课人数',
+                xtype: 'numberfield',
+                minValue: 0,
                 name: 'quantity'
             }, {
                 fieldLabel: '多媒体设备',
@@ -163,12 +171,12 @@ courseManage = {
                 course: value
             },
             success: function(){
-                Ext.Msg.alert('提示', "创建成功");
+                Wando.msg.info('提示', "创建成功");
                 Ext.getCmp('courseForm').getForm().reset();
                 Ext.getCmp('courseGrid').store.reload();
             },
             failure: function(){
-                Ext.Msg.alert('提示', "创建失败");
+                Wando.msg.error('提示', "创建失败");
             }
         })
     },
@@ -183,11 +191,11 @@ courseManage = {
                 id: sel.data.id
             },
             success: function(){
-                Ext.Msg.alert('提示', "删除成功");
+                Wando.msg.info('提示', "删除成功");
                 grid.store.remove(sel);
             },
             failure: function(){
-                Ext.Msg.alert('提示', "删除失败");
+                Wando.msg.error('提示', "删除失败");
             }
         })
     }
